pyproject.toml
[project]
name = "youarebot"
version = "0.1.0"
description = ""
authors = ""
requires-python = "~=3.12.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "numpy>=1.26.4,<2",
    "PyYAML",
    "boto3",
    "tqdm",
    "clickhouse-driver",
    "ipykernel",
    "minio",
    "pandas",
    "psycopg2-binary",
    "catboost",
    "python-dotenv",
    "scikit-learn",
    "pyarrow",
    "requests",
    "mlflow",
    "sqlalchemy",
    "black",
    "booktype",
    "schemas",
    "redis",
    "aiohttp>=3.13.1",
    "skl2onnx>=1.19.1",
    "onnxruntime>=1.23.1",
    "onnx>=1.19.1",
    "onnxscript>=0.5.4",
    "ort>=1.2.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1,<2",
    "ruff>=0.8.4,<0.9",
    "ipdb>=0.13.13,<0.14",
    "isort>=5.13.2,<6",
    "mypy>=1.13.0,<2",
    "pre-commit>=4.0.1,<5"
]

[tool.uv]
package = false

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages"
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D1", "COM812", "TRY003", "EM101"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["S101", "ANN401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
